Correct the globule size and position
Annoyingly, these need to be given in terms of the physical grid size
in the y direction.  Now that I have turned the grid round (to conform
to the code's expectations about which axis is the cylindrical axis),
the y size is 4 times larger. So all lengths need to be reduced by a
factor of four